refactor(files): 文件展示统一为 fileview/ 模块 + 各类型子组件
把原先塞在 FileBrowser.tsx 里的巨型 Viewer(~300 行,图片/PDF/Office/CSV/
markdown/HTML/代码全在一个三元里)拆成层层封装的组件树：

- src/file-utils.ts   中性纯工具:ext/大小/路径/lang 映射/CSV 解析 + fileKind() 分类
- src/file-icons.tsx  中性图标:FileTypeIcon + 导航图标 + IconButton/ClosePanelButton
- src/fileview/
    FileView.tsx      容器:取数/活重载/脏保存/源码⇄渲染/标题栏/inline·modal 外壳
    PreviewShell.tsx  带边框预览外壳(PDF/Office/CSV 共用)
    ImageView / PdfView / HtmlView / MarkdownView / CsvView / CodeView / OfficeView
    index.ts          对外导出 FileView(含 Viewer 兼容别名)

依赖为无环 DAG:file-utils ← file-icons ← fileview/* ← FileBrowser ← FileWorkspace。
页面「展示一个文件」统一 import { FileView } from './fileview'。加新类型只需加一个
子组件 + 在容器分发里挂一支。

- FileBrowser.tsx 1451→844 行,只保留「目录浏览器」职责,展示交给 FileView。
- FileView 为原 Viewer 的逐行行为保留移植(props/取数/活重载/脏保存/HTML 新标签
  按钮/floating 钮/stale 提示全部照旧),对外改名 FileView 并保留 Viewer 别名。
- 懒加载分块(monaco/office/markdown)保留,主包体积不变。
