+++ frontend/src/FileBrowser.tsx
@@ -1,42 +1,16 @@
 // 文件侧栏 —— 在 Claude / Codex 对话页右侧浏览工作目录、查看文件内容（类似 codex 右侧边栏）。
 // 单层可导航列表：目录在前可进入、↑ 回上级、点文件在弹层里查看正文。
-import { type MouseEvent, type ReactNode, Fragment, lazy, Suspense, useEffect, useMemo, useRef, useState } from 'react'
-import { AutoComplete, Button, ConfigProvider, Dropdown, Input, Modal, Space, Spin, App as AntApp, Tooltip, type MenuProps } from 'antd'
+import { type ReactNode, Fragment, useEffect, useMemo, useRef, useState } from 'react'
+import { AutoComplete, Button, ConfigProvider, Dropdown, Input, Modal, Spin, App as AntApp, Tooltip, type MenuProps } from 'antd'
 import { api, upload } from './api'
-import Markdown from './Markdown'
-import ErrorBoundary from './ErrorBoundary'
 import { useI18n } from './i18n'
 import { recentDirs } from './App'
-import { useThemeMode } from './theme'
-
-// Monaco 代码编辑器很重，只有真正编辑/查看文本文件才需要 → 懒加载，不进首屏包。
-const CodeEditor = lazy(() => import('./CodeEditor'))
-
-// Office 预览（docx-preview / xlsx / pptx）依赖很重，只有真正打开 Office 文件才需要：
-// 懒加载使其不进入首屏包，按需异步取。
-const OfficePreviewers = () => import('./OfficePreviewers')
-const DocxFilePreview = lazy(() => OfficePreviewers().then((m) => ({ default: m.DocxFilePreview })))
-const ExcelFilePreview = lazy(() => OfficePreviewers().then((m) => ({ default: m.ExcelFilePreview })))
-const PptxFilePreview = lazy(() => OfficePreviewers().then((m) => ({ default: m.PptxFilePreview })))
-
-const IMG_EXT = ['png', 'jpg', 'jpeg', 'gif', 'webp', 'bmp', 'ico', 'avif', 'svg']
-const MD_EXT = ['md', 'markdown', 'mdx']
-const CODE_LANG: Record<string, string> = {
-  py: 'python', pyw: 'python', sh: 'bash', bash: 'bash', zsh: 'bash', fish: 'bash', ps1: 'powershell',
-  js: 'javascript', jsx: 'jsx', ts: 'typescript', tsx: 'tsx', mjs: 'javascript', cjs: 'javascript',
-  go: 'go', rs: 'rust', java: 'java', kt: 'kotlin', c: 'c', h: 'c', cpp: 'cpp', hpp: 'cpp',
-  cs: 'csharp', php: 'php', rb: 'ruby', swift: 'swift', html: 'html', htm: 'html', css: 'css',
-  scss: 'scss', sass: 'sass', less: 'less', sql: 'sql', json: 'json', yaml: 'yaml', yml: 'yaml',
-  toml: 'toml', xml: 'xml', ini: 'ini', env: 'ini', conf: 'ini', Dockerfile: 'dockerfile',
-}
-function extOf(path: string): string {
-  const m = path.toLowerCase().match(/\.([a-z0-9]+)$/)
-  return m ? m[1] : ''
-}
-
-function fileNameOf(path: string): string {
-  return path.split('/').pop() || 'download'
-}
+import { dirname, fileNameOf, fmtSize, joinPath, normalizePath } from './file-utils'
+import {
+  BackIcon, Chevron, ClosePanelButton, CloseIcon, DownloadIcon, EyeIcon, EyeOffIcon, FileTypeIcon, FolderIcon,
+  FolderUpIcon, ForwardIcon, IconButton, ListIcon, NewFolderIcon, RefreshIcon, SearchIcon, SortIcon, TreeIcon, UploadIcon,
+} from './file-icons'
+import { Viewer } from './fileview'
 
 interface Entry { name: string; dir: boolean; size: number; mtime: number; ctime: number }
 interface Dir { path: string; parent: string; entries: Entry[] }
@@ -75,152 +49,10 @@ function sortEntries(entries: Entry[], key: SortKey): Entry[] {
   return sorted
 }
 
-function fmtSize(n: number): string {
-  if (n < 1024) return n + ' B'
-  if (n < 1024 * 1024) return (n / 1024).toFixed(1) + ' K'
-  return (n / 1024 / 1024).toFixed(1) + ' M'
-}
-
-function joinPath(dir: string, name: string): string {
-  return (dir === '/' ? '' : dir) + '/' + name
-}
-
-function dirname(path: string): string {
-  const i = path.lastIndexOf('/')
-  return i <= 0 ? '/' : path.slice(0, i)
-}
-
-function normalizePath(path: string): string {
-  const abs = path.startsWith('/')
-  const parts: string[] = []
-  for (const part of path.split('/')) {
-    if (!part || part === '.') continue
-    if (part === '..') parts.pop()
-    else parts.push(part)
-  }
-  return (abs ? '/' : '') + parts.join('/')
-}
-
 function displayPath(path: string): string {
   return path || '/'
 }
 
-function stripHashQuery(ref: string): string {
-  return ref.split('#')[0].split('?')[0]
-}
-
-function localPathFromRef(baseFile: string, ref: string): string | null {
-  const raw = stripHashQuery(ref.trim())
-  if (!raw || raw.startsWith('#') || raw.startsWith('//') || /^[a-z][a-z0-9+.-]*:/i.test(raw)) return null
-  let clean = raw
-  try { clean = decodeURIComponent(raw) } catch { /* keep raw */ }
-  return normalizePath(clean.startsWith('/') ? clean : joinPath(dirname(baseFile), clean))
-}
-
-function codeLangOf(path: string): string {
-  const name = path.split('/').pop() || ''
-  if (name === 'Dockerfile' || name.endsWith('.Dockerfile')) return 'dockerfile'
-  return CODE_LANG[extOf(path)] || ''
-}
-
-// Monaco 语言 id（与 highlight.js 的略有出入：bash→shell、tsx→typescript、toml→ini…）
-const MONACO_LANG: Record<string, string> = {
-  py: 'python', pyw: 'python', ipynb: 'json', sh: 'shell', bash: 'shell', zsh: 'shell', fish: 'shell', ps1: 'powershell',
-  js: 'javascript', jsx: 'javascript', mjs: 'javascript', cjs: 'javascript', ts: 'typescript', tsx: 'typescript',
-  go: 'go', rs: 'rust', java: 'java', kt: 'kotlin', c: 'c', h: 'c', cpp: 'cpp', hpp: 'cpp', cs: 'csharp',
-  php: 'php', rb: 'ruby', swift: 'swift', html: 'html', htm: 'html', vue: 'html', css: 'css', scss: 'scss', sass: 'scss',
-  less: 'less', sql: 'sql', json: 'json', jsonl: 'json', ndjson: 'json', yaml: 'yaml', yml: 'yaml',
-  toml: 'ini', ini: 'ini', env: 'ini', conf: 'ini', lock: 'ini', xml: 'xml', md: 'markdown', markdown: 'markdown', mdx: 'markdown',
-}
-function monacoLangOf(path: string): string {
-  const name = path.split('/').pop() || ''
-  if (name === 'Dockerfile' || name.endsWith('.Dockerfile')) return 'dockerfile'
-  return MONACO_LANG[extOf(path)] || 'plaintext'
-}
-
-function parseDelimited(text: string, sep: ',' | '\t'): string[][] {
-  const rows: string[][] = []
-  let row: string[] = [], cell = '', quote = false
-  for (let i = 0; i < text.length; i++) {
-    const ch = text[i]
-    if (quote && ch === '"' && text[i + 1] === '"') { cell += '"'; i++; continue }
-    if (ch === '"') { quote = !quote; continue }
-    if (!quote && ch === sep) { row.push(cell); cell = ''; continue }
-    if (!quote && (ch === '\n' || ch === '\r')) {
-      if (ch === '\r' && text[i + 1] === '\n') i++
-      row.push(cell); rows.push(row); row = []; cell = ''
-      continue
-    }
-    cell += ch
-  }
-  if (cell || row.length) { row.push(cell); rows.push(row) }
-  return rows.filter((r) => r.some((x) => x.trim() !== '')).slice(0, 80).map((r) => r.slice(0, 12))
-}
-
-// 目录/文件图标：按 blade-agent 的分类 SVG 思路做轻量映射，避免字母块图标。
-const FolderIcon = () => (
-  <svg width="18" height="18" viewBox="0 0 24 24" fill="currentColor" opacity="0.9"><path d="M20 6h-8l-2-2H4a2 2 0 0 0-2 2v12a2 2 0 0 0 2 2h16a2 2 0 0 0 2-2V8a2 2 0 0 0-2-2z" /></svg>
-)
-const FileIcon = () => (
-  <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.9" strokeLinecap="round" strokeLinejoin="round"><path d="M14.5 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V7.5L14.5 2z" /><polyline points="14 2 14 8 20 8" /></svg>
-)
-const CodeIcon = () => <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.9" strokeLinecap="round" strokeLinejoin="round"><polyline points="16 18 22 12 16 6" /><polyline points="8 6 2 12 8 18" /></svg>
-const TableIcon = () => <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.9" strokeLinecap="round" strokeLinejoin="round"><rect x="3" y="3" width="18" height="18" rx="2" /><path d="M3 9h18" /><path d="M3 15h18" /><path d="M9 3v18" /></svg>
-const SlidesIcon = () => <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.9" strokeLinecap="round" strokeLinejoin="round"><rect x="2" y="3" width="20" height="14" rx="2" /><path d="M8 21h8" /><path d="M12 17v4" /></svg>
-const PdfIcon = () => <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.9" strokeLinecap="round" strokeLinejoin="round"><path d="M14.5 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V7.5L14.5 2z" /><path d="M8 13h8" /><path d="M8 17h5" /></svg>
-const ImageIcon = () => <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.9" strokeLinecap="round" strokeLinejoin="round"><rect x="3" y="3" width="18" height="18" rx="2" /><circle cx="9" cy="9" r="2" /><path d="m21 15-3.09-3.09a2 2 0 0 0-2.82 0L6 21" /></svg>
-const TextIcon = () => <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.9" strokeLinecap="round" strokeLinejoin="round"><path d="M14.5 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V7.5L14.5 2z" /><path d="M8 13h8" /><path d="M8 17h8" /><path d="M8 9h3" /></svg>
-const ArchiveIcon = () => <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.9" strokeLinecap="round" strokeLinejoin="round"><rect x="2" y="4" width="20" height="16" rx="2" /><path d="M12 4v16" /><path d="m10 10 4 4" /><path d="m14 10-4 4" /></svg>
-const ConfigIcon = () => <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.9" strokeLinecap="round" strokeLinejoin="round"><path d="M14.5 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V7.5L14.5 2z" /><circle cx="12" cy="15" r="2" /><path d="M12 11v2" /></svg>
-const DatabaseIcon = () => <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.9" strokeLinecap="round" strokeLinejoin="round"><ellipse cx="12" cy="5" rx="9" ry="3" /><path d="M3 5v14a9 3 0 0 0 18 0V5" /><path d="M3 12a9 3 0 0 0 18 0" /></svg>
-const AudioIcon = () => <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.9" strokeLinecap="round" strokeLinejoin="round"><path d="M9 18V5l12-2v13" /><circle cx="6" cy="18" r="3" /><circle cx="18" cy="16" r="3" /></svg>
-const VideoIcon = () => <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.9" strokeLinecap="round" strokeLinejoin="round"><rect x="2" y="2" width="20" height="20" rx="2" /><path d="m10 8 6 4-6 4V8z" /></svg>
-const DesignIcon = () => <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.9" strokeLinecap="round" strokeLinejoin="round"><path d="m12 19 7-7 3 3-7 7-3-3z" /><path d="m18 13-1.5-7.5L2 2l3.5 14.5L13 18l5-5z" /><path d="m2 2 7.586 7.586" /><circle cx="11" cy="11" r="2" /></svg>
-type FileIconEntry = { icon: ReactNode; color: string }
-const EXT_ICON: Record<string, FileIconEntry> = {}
-function regIcon(exts: string[], icon: ReactNode, color: string) {
-  for (const ext of exts) EXT_ICON[ext] = { icon, color }
-}
-regIcon(['py', 'pyw', 'ipynb'], <CodeIcon />, 'hsl(210,70%,58%)')
-regIcon(['js', 'jsx', 'mjs', 'cjs'], <CodeIcon />, 'hsl(48,90%,50%)')
-regIcon(['ts', 'tsx'], <CodeIcon />, 'hsl(210,80%,55%)')
-regIcon(['go'], <CodeIcon />, 'hsl(195,70%,50%)')
-regIcon(['rs'], <CodeIcon />, 'hsl(25,80%,55%)')
-regIcon(['java', 'kt'], <CodeIcon />, 'hsl(20,85%,52%)')
-regIcon(['rb'], <CodeIcon />, 'hsl(0,70%,55%)')
-regIcon(['c', 'cpp', 'h', 'hpp'], <CodeIcon />, 'hsl(210,60%,55%)')
-regIcon(['cs'], <CodeIcon />, 'hsl(265,55%,55%)')
-regIcon(['swift'], <CodeIcon />, 'hsl(20,90%,55%)')
-regIcon(['html', 'htm'], <CodeIcon />, 'hsl(15,85%,55%)')
-regIcon(['css', 'scss', 'sass', 'less'], <CodeIcon />, 'hsl(210,70%,55%)')
-regIcon(['vue'], <CodeIcon />, 'hsl(153,60%,48%)')
-regIcon(['php'], <CodeIcon />, 'hsl(240,40%,58%)')
-regIcon(['sh', 'bash', 'zsh', 'fish', 'ps1'], <CodeIcon />, 'var(--text-dim)')
-regIcon(['sql'], <CodeIcon />, 'hsl(210,50%,55%)')
-regIcon(['json', 'jsonl', 'ndjson'], <ConfigIcon />, 'hsl(158,55%,48%)')
-regIcon(['yaml', 'yml', 'toml', 'ini', 'conf', 'env', 'lock'], <ConfigIcon />, 'var(--text-dim)')
-regIcon(['xml'], <CodeIcon />, 'hsl(25,65%,52%)')
-regIcon(['db', 'sqlite', 'parquet'], <DatabaseIcon />, 'hsl(210,50%,55%)')
-regIcon(['doc', 'docx', 'odt', 'rtf', 'pages'], <TextIcon />, 'hsl(210,65%,52%)')
-regIcon(['xls', 'xlsx', 'xlsm', 'csv', 'tsv', 'ods', 'numbers'], <TableIcon />, 'hsl(140,55%,42%)')
-regIcon(['ppt', 'pptx', 'odp', 'key'], <SlidesIcon />, 'hsl(15,80%,52%)')
-regIcon(['pdf'], <PdfIcon />, 'hsl(0,65%,50%)')
-regIcon(['md', 'markdown', 'txt', 'log', 'tex', 'epub'], <TextIcon />, 'var(--text-dim)')
-regIcon(['png', 'jpg', 'jpeg', 'gif', 'webp', 'bmp', 'ico', 'heic', 'heif', 'avif', 'tif', 'tiff', 'svg'], <ImageIcon />, 'hsl(280,55%,58%)')
-regIcon(['psd', 'ai', 'fig', 'sketch', 'xd', 'blend'], <DesignIcon />, 'hsl(280,50%,55%)')
-regIcon(['mp3', 'wav', 'flac', 'ogg', 'aac', 'aiff', 'm4a', 'mid'], <AudioIcon />, 'hsl(330,60%,55%)')
-regIcon(['mp4', 'mov', 'mkv', 'avi', 'webm', 'wmv', 'flv', 'm4v'], <VideoIcon />, 'hsl(340,65%,52%)')
-regIcon(['zip', 'tar', 'gz', 'rar', '7z', 'bz2', 'xz', 'tgz', 'dmg', 'iso', 'pkg'], <ArchiveIcon />, 'hsl(30,50%,48%)')
-regIcon(['ttf', 'otf', 'woff', 'woff2'], <FileIcon />, 'var(--text-dim)')
-export const FileTypeIcon = ({ name }: { name: string }) => {
-  const entry = EXT_ICON[extOf(name)] || { icon: <FileIcon />, color: 'var(--text-dimmer)' }
-  return (
-    <span style={{
-      width: 22, height: 22, display: 'inline-grid', placeItems: 'center', color: entry.color,
-      flex: '0 0 auto',
-    }}>{entry.icon}</span>
-  )
-}
 const PathOption = ({ kind, path, name, dir }: { kind: string; path: string; name?: string; dir?: boolean }) => (
   <div style={{ display: 'flex', alignItems: 'center', gap: 8, minWidth: 260, maxWidth: 560 }}>
     <span style={{ color: dir ? 'var(--text-bright)' : 'var(--text-dimmer)', width: 20, display: 'inline-flex', justifyContent: 'center' }}>
@@ -232,442 +64,6 @@ const PathOption = ({ kind, path, name, dir }: { kind: string; path: string; nam
     <span style={{ color: 'var(--text-dimmer)', fontSize: 11, flex: '0 0 auto' }}>{kind}</span>
   </div>
 )
-const FolderUpIcon = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round"><path d="M3 7a2 2 0 0 1 2-2h4l2 2h8a2 2 0 0 1 2 2v8a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2z" /><path d="M12 16V10" /><path d="m9 13 3-3 3 3" /></svg>
-)
-const RefreshIcon = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round"><path d="M21 12a9 9 0 0 1-15 6.7" /><path d="M3 12A9 9 0 0 1 18 5.3" /><path d="M18 2v4h-4" /><path d="M6 22v-4h4" /></svg>
-)
-const BackIcon = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.9" strokeLinecap="round" strokeLinejoin="round"><path d="m15 18-6-6 6-6" /></svg>
-)
-const ForwardIcon = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.9" strokeLinecap="round" strokeLinejoin="round"><path d="m9 18 6-6-6-6" /></svg>
-)
-const NewFolderIcon = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round"><path d="M20 20a2 2 0 0 0 2-2V8a2 2 0 0 0-2-2h-7.9a2 2 0 0 1-1.69-.9L9.6 3.9A2 2 0 0 0 7.93 3H4a2 2 0 0 0-2 2v13a2 2 0 0 0 2 2Z" /><path d="M12 10v6" /><path d="M9 13h6" /></svg>
-)
-const UploadIcon = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round"><path d="M16 16h2a4 4 0 0 0 .7-7.9A6 6 0 0 0 7.4 6.8 5 5 0 0 0 8 16h1" /><path d="M12 18V10" /><path d="m8.5 13.5 3.5-3.5 3.5 3.5" /></svg>
-)
-const DownloadIcon = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round"><path d="M12 3v10" /><path d="m8.5 9.5 3.5 3.5 3.5-3.5" /><path d="M4 17v2a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2v-2" /></svg>
-)
-const CloseIcon = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round"><path d="M18 6 6 18" /><path d="m6 6 12 12" /></svg>
-)
-const EyeIcon = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round"><path d="M2 12s3.5-7 10-7 10 7 10 7-3.5 7-10 7-10-7-10-7Z" /><circle cx="12" cy="12" r="3" /></svg>
-)
-const EyeOffIcon = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round"><path d="M9.9 4.2A10.9 10.9 0 0 1 12 4c6.5 0 10 7 10 7a17.7 17.7 0 0 1-3 3.7" /><path d="M6.6 6.6A17.6 17.6 0 0 0 2 11s3.5 7 10 7a10.6 10.6 0 0 0 4.4-.9" /><path d="m2 2 20 20" /><path d="M9.5 9.5a3 3 0 0 0 4.2 4.2" /></svg>
-)
-const SortIcon = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round"><path d="M3 6h7" /><path d="M3 12h10" /><path d="M3 18h14" /><path d="M18 6v12" /><path d="m15 15 3 3 3-3" /></svg>
-)
-// 树形视图开关：平铺列表 <-> 可展开目录树
-const TreeIcon = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round"><rect x="3" y="3" width="6" height="6" rx="1" /><rect x="15" y="15" width="6" height="6" rx="1" /><path d="M9 6h6a2 2 0 0 1 2 2v7" /></svg>
-)
-const ListIcon = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round"><path d="M8 6h13" /><path d="M8 12h13" /><path d="M8 18h13" /><path d="M3 6h.01" /><path d="M3 12h.01" /><path d="M3 18h.01" /></svg>
-)
-const SearchIcon = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round"><circle cx="11" cy="11" r="7" /><path d="m21 21-4.3-4.3" /></svg>
-)
-// markdown 预览（VSCode 式）：切换预览 / 侧栏打开预览
-const PreviewIcon = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.7" strokeLinecap="round" strokeLinejoin="round"><rect x="3" y="4" width="18" height="16" rx="2" /><path d="M7 9h10" /><path d="M7 13h10" /><path d="M7 17h6" /></svg>
-)
-const PreviewSideIcon = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.7" strokeLinecap="round" strokeLinejoin="round"><rect x="3" y="4" width="18" height="16" rx="2" /><path d="M13 4v16" /><path d="M16 10h3" /><path d="M16 14h3" /></svg>
-)
-const ExternalLinkIcon = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.7" strokeLinecap="round" strokeLinejoin="round"><path d="M15 3h6v6" /><path d="M10 14 21 3" /><path d="M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6" /></svg>
-)
-// 目录展开箭头：展开时旋转 90°
-const Chevron = ({ open }: { open: boolean }) => (
-  <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.2" strokeLinecap="round" strokeLinejoin="round"
-    style={{ transition: 'transform .12s', transform: open ? 'rotate(90deg)' : 'none' }}><path d="m9 18 6-6-6-6" /></svg>
-)
-const IconButton = ({ title, children, danger, onClick, disabled, width = 24 }: { title: string; children: React.ReactNode; danger?: boolean; onClick?: (e: React.MouseEvent) => void; disabled?: boolean; width?: number | string }) => (
-  <Tooltip title={title}>
-    <Button type="text" size="small" disabled={disabled} danger={danger} onClick={(e) => { e.stopPropagation(); onClick?.(e) }}
-      style={{ width, height: 24, minWidth: 24, padding: width === 24 ? 0 : '0 6px', display: 'inline-flex', alignItems: 'center', justifyContent: 'center', gap: 4 }}>
-      {children}
-    </Button>
-  </Tooltip>
-)
-
-const ClosePanelButton = ({ title, onClick }: { title: string; onClick: () => void }) => (
-  <button
-    type="button"
-    title={title}
-    aria-label={title}
-    className="tt-file-close"
-    onMouseDown={(e) => { e.preventDefault(); e.stopPropagation() }}
-    onClick={(e) => { e.preventDefault(); e.stopPropagation(); onClick() }}
-  >
-    <CloseIcon />
-  </button>
-)
-
-function OfficePreview({ name, previewUrl, rawUrl, downloadUrl, downloadName, height }: { name: string; previewUrl: string; rawUrl: string; downloadUrl: string; downloadName: string; height: string }) {
-  const { t } = useI18n()
-  const [url, setUrl] = useState('')
-  const [err, setErr] = useState('')
-  useEffect(() => {
-    let stop = false
-    let objectUrl = ''
-    setUrl('')
-    setErr('')
-    fetch(previewUrl).then(async (r) => {
-      if (!r.ok) {
-        const data = await r.json().catch(() => null)
-        throw new Error(data?.error?.message || data?.error?.code || `HTTP ${r.status}`)
-      }
-      return r.blob()
-    }).then((blob) => {
-      if (stop) return
-      objectUrl = URL.createObjectURL(blob)
-      setUrl(objectUrl)
-    }).catch((e) => {
-      if (!stop) setErr(e.message || 'Office 预览生成失败')
-    })
-    return () => {
-      stop = true
-      if (objectUrl) URL.revokeObjectURL(objectUrl)
-    }
-  }, [previewUrl])
-
-  if (err) {
-    return (
-      <div style={{ minHeight: height, padding: 18, color: 'var(--text-dim)', lineHeight: 1.7 }}>
-        <div style={{ color: 'var(--text-bright)', fontWeight: 700, marginBottom: 6 }}>{name}</div>
-        <div>{err}</div>
-        <div style={{ marginTop: 14, display: 'flex', gap: 8, flexWrap: 'wrap' }}>
-          <Button size="small" type="primary" href={downloadUrl} download={downloadName}>{t('file.downloadFile')}</Button>
-          <Button size="small" href={rawUrl} target="_blank">{t('file.openRaw')}</Button>
-        </div>
-      </div>
-    )
-  }
-  if (!url) return <div style={{ height, display: 'grid', placeItems: 'center' }}><Spin /></div>
-  return <iframe title={name} src={url} style={{ width: '100%', height, border: 0, background: '#fff' }} />
-}
-
-export function Viewer({
-  path,
-  accent,
-  inline,
-  active = true,
-  tabbed,
-  forcePreview,
-  onClose,
-  onOpenPath,
-  onOpenAgent,
-  onDirtyChange,
-  onPreviewToSide,
-}: {
-  path: string
-  accent: string
-  inline?: boolean
-  // 多 tab 常驻挂载时，只有激活的才渲染重型 Monaco 编辑器/预览，避免多实例吃爆内存(OOM)。
-  // 组件仍挂载 → draft/dirty 等 state 保留，切回来编辑内容不丢。
-  active?: boolean
-  // 编辑器 tab 上下文：外层 tab 已显示文件名+关闭，这里就不再重复顶部「▸ 文件名」标题行。
-  tabbed?: boolean
-  // 专用预览 tab（VSCode「侧栏预览」）：始终渲染 markdown，不显示编辑器/切换钮。
-  forcePreview?: boolean
-  onClose: () => void
-  onOpenPath: (p: string) => void
-  onOpenAgent?: (kind: 'claude' | 'codex', path: string) => void
-  // 编辑器脏状态上报（供外层 tab 显示未保存圆点）
-  onDirtyChange?: (path: string, dirty: boolean) => void
-  // 在侧栏打开渲染预览（外层 FileWorkspace 处理，开另一栏）
-  onPreviewToSide?: (path: string) => void
-}) {
-  const ext = extOf(path)
-  const isImg = IMG_EXT.includes(ext)
-  const isMd = MD_EXT.includes(ext)
-  const isHtml = ext === 'html' || ext === 'htm'
-  const isPdf = ext === 'pdf'
-  const isOffice = ['doc', 'docx', 'odt', 'rtf', 'xls', 'xlsx', 'xlsm', 'ods', 'ppt', 'pptx', 'odp'].includes(ext)
-  const isDocxPreview = ext === 'docx'
-  const isExcelPreview = ['xls', 'xlsx', 'xlsm', 'ods'].includes(ext)
-  const isPptxPreview = ext === 'pptx'
-  const isSheetText = ['csv', 'tsv'].includes(ext)
-  const codeLang = codeLangOf(path)
-  const rawUrl = `/api/file/raw?path=${encodeURIComponent(path)}`
-  const previewUrl = `/api/file/preview?path=${encodeURIComponent(path)}`
-  const downloadUrl = `${rawUrl}&dl=1`
-  const downloadName = fileNameOf(path)
-  const previewHeight = inline ? '100%' : '74vh'
-  const [data, setData] = useState<any>(null)
-  const [err, setErr] = useState('')
-  const [source, setSource] = useState(false) // markdown：源码/渲染切换
-  const [agentPick, setAgentPick] = useState(false)
-  const [draft, setDraft] = useState('') // 编辑器当前文本
-  const [saving, setSaving] = useState(false)
-  const [stale, setStale] = useState(false) // 磁盘上文件已被外部(cc/codex)改动，但本地有未保存改动没自动覆盖
-  const { message } = AntApp.useApp()
-  const { t } = useI18n()
-  const { mode } = useThemeMode()
-
-  // 可编辑：文本/代码/JSON/Markdown（源码）；二进制、被截断的大文件、表格/图片/PDF/Office 不可编辑。
-  const editable = !!data && !data.binary && !data.truncated && !isSheetText && !isImg && !isPdf && !isOffice
-  const dirty = editable && data ? draft !== data.content : false
-  useEffect(() => { onDirtyChange?.(path, dirty); return () => onDirtyChange?.(path, false) }, [dirty, path])
-
-  const save = async () => {
-    if (!editable || saving || !dirty) return
-    setSaving(true)
-    try {
-      const res = await api('POST', '/file/save', { path, content: draft })
-      setData((d: any) => ({ ...d, content: draft, mtime: res.data?.mtime ?? d?.mtime })) // 基线更新 → dirty 归零；mtime 同步避免自触发重载
-      setStale(false)
-      message.success(t('file.saved'))
-    } catch (e: any) {
-      message.error(t('file.saveFailed', { message: e.message }))
-    } finally {
-      setSaving(false)
-    }
-  }
-  // Monaco 的 Ctrl+S 命令在挂载时捕获闭包，用 ref 始终指向最新 save，避免存到旧文本。
-  const saveRef = useRef(save)
-  saveRef.current = save
-
-  useEffect(() => {
-    if (isImg || isPdf || isOffice) return // 图片/PDF/Office 直接走 raw 或专用面板
-    // tab 语境的 markdown 默认进编辑器（源码），点眼睛才切预览；非 tab（有头部）默认渲染。
-    // tab 语境的 markdown 默认进编辑器(源码)、点预览才渲染;HTML 则默认直接渲染
-    // (打开网页多是想看效果,不是编辑),两者都可用「源码/渲染」钮切换。
-    setData(null); setErr(''); setStale(false); setSource(!!tabbed && MD_EXT.includes(extOf(path))); setDraft('')
-    api('GET', `/file?path=${encodeURIComponent(path)}`).then((r) => { setData(r.data); setDraft(r.data?.content || '') }).catch((e) => setErr(e.message))
-  }, [path, isImg, isPdf, isOffice])
-
-  // 从磁盘重载（放弃本地未保存改动）
-  const reloadFromDisk = () => {
-    api('GET', `/file?path=${encodeURIComponent(path)}`).then((r) => { setData(r.data); setDraft(r.data?.content || ''); setStale(false) }).catch(() => {})
-  }
-  // 外部(cc/codex 等)改动已打开的文件 → 轮询 mtime：无本地改动自动重载渲染；有未保存改动只提示不覆盖。
-  useEffect(() => {
-    if (!active || !data || err || isImg || isPdf || isOffice) return
-    let stop = false
-    const h = setInterval(async () => {
-      try {
-        const r = await api('GET', `/file/stat?path=${encodeURIComponent(path)}`)
-        if (stop || !r.data?.mtime || r.data.mtime === data.mtime) return
-        if (dirty) { setStale(true); return } // 有未保存改动 → 不覆盖，仅提示
-        const fr = await api('GET', `/file?path=${encodeURIComponent(path)}`)
-        if (!stop) { setData(fr.data); setDraft(fr.data?.content || '') }
-      } catch {}
-    }, 2000)
-    return () => { stop = true; clearInterval(h) }
-  }, [active, data?.mtime, dirty, path, isImg, isPdf, isOffice, err])
-
-  // 非激活 tab：只占位、不挂载重型 Monaco/预览（state 已在上面 hook 里保留，切回来不丢编辑）。
-  if (!active) return <div style={{ height: '100%' }} />
-
-  const name = fileNameOf(path)
-  const copyPath = async () => {
-    try {
-      await navigator.clipboard.writeText(path)
-      message.success(t('file.pathCopied'))
-    } catch {
-      message.error(t('common.copyFailed'))
-    }
-  }
-  const codePre = (text: string) => (
-    <pre style={{ margin: 0, whiteSpace: 'pre', overflow: 'auto', height: previewHeight, background: 'var(--bg-base)', padding: 12, borderRadius: 8, fontFamily: 'ui-monospace, monospace', fontSize: 12.5, lineHeight: 1.5, color: 'var(--text-bright)' }}>{text}</pre>
-  )
-  const resolvePreviewHref = (href: string, kind: 'link' | 'image') => {
-    const local = localPathFromRef(path, href)
-    if (!local) return href
-    if (kind === 'image') return `/api/file/raw?path=${encodeURIComponent(local)}`
-    return `/api/file/raw?path=${encodeURIComponent(local)}`
-  }
-  const openPreviewLink = (href: string, ev: MouseEvent<HTMLAnchorElement>) => {
-    const local = localPathFromRef(path, href)
-    if (!local) return
-    ev.preventDefault()
-    onOpenPath(local)
-  }
-  const previewShell = (title: string, body: React.ReactNode) => (
-    <div style={{ border: '1px solid var(--border-subtle)', borderRadius: 8, overflow: 'hidden', background: 'var(--bg-base)', height: inline ? '100%' : undefined, display: 'flex', flexDirection: 'column' }}>
-      <div style={{ display: 'flex', alignItems: 'center', gap: 8, padding: '8px 10px', borderBottom: '1px solid var(--border-subtle)', color: 'var(--text-dim)', fontSize: 12 }}>
-        <FileTypeIcon name={path} />
-        <span>{title}</span>
-        {data?.truncated && <span style={{ marginLeft: 'auto', color: '#d29922' }}>{t('file.truncatedShort')}</span>}
-      </div>
-      <div style={{ flex: 1, minHeight: 0, overflow: 'hidden' }}>{body}</div>
-    </div>
-  )
-  const csvTable = (text: string, sep: ',' | '\t') => {
-    const rows = parseDelimited(text, sep)
-    const head = rows[0] || []
-    const body = rows.slice(1)
-    return previewShell(t('file.tablePreviewTitle', { kind: sep === ',' ? 'CSV' : 'TSV' }), (
-      <div style={{ height: previewHeight, overflow: 'auto' }}>
-        <table style={{ width: '100%', borderCollapse: 'collapse', fontSize: 12 }}>
-          <thead>{head.length > 0 && <tr>{head.map((c, i) => <th key={i} style={cellStyle(true)}>{c || t('file.column', { index: i + 1 })}</th>)}</tr>}</thead>
-          <tbody>{body.map((r, i) => <tr key={i}>{head.map((_, j) => <td key={j} style={cellStyle(false)}>{r[j] || ''}</td>)}</tr>)}</tbody>
-        </table>
-      </div>
-    ))
-  }
-
-  const titleNode = (
-    <div style={{ display: 'flex', alignItems: 'center', gap: 10, paddingRight: inline ? 0 : 28, minWidth: 0 }}>
-      {!tabbed && (
-        <span style={{ fontFamily: 'ui-monospace, monospace', fontSize: 13, overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap' }}>
-          <span style={{ color: accent }}>▸</span> {name}
-        </span>
-      )}
-      {dirty && <span title={t('file.unsaved')} style={{ width: 8, height: 8, borderRadius: '50%', background: '#d29922', flex: '0 0 auto' }} />}
-      <span style={{ flex: 1, minWidth: 8 }} />
-      {editable && (
-        <Button size="small" type="primary" ghost={!dirty} disabled={!dirty || saving} loading={saving} onClick={save}>{t('file.save')}</Button>
-      )}
-      {(isMd || isHtml) && data && !data.binary && (
-        <Button size="small" onClick={() => setSource((s) => !s)}>{source ? t('file.rendered') : t('file.source')}</Button>
-      )}
-      {isHtml && (
-        <Button size="small" href={rawUrl} target="_blank" rel="noreferrer">{t('file.openInNewTab')}</Button>
-      )}
-      {onOpenAgent && (
-        <Button size="small" onClick={() => setAgentPick(true)}>{t('file.openInAgent')}</Button>
-      )}
-      <Button size="small" onClick={copyPath}>{t('file.copyPath')}</Button>
-      <Button size="small" href={downloadUrl} download={downloadName}>{t('file.download')}</Button>
-      <a href={rawUrl} target="_blank" rel="noreferrer" style={{ color: 'var(--text-dim)', fontSize: 12 }}>{t('file.raw')}</a>
-      {inline && !tabbed && <IconButton title={t('file.closePreview')} onClick={onClose}><CloseIcon /></IconButton>}
-    </div>
-  )
-  const bodyNode = (
-    <>
-      {isImg ? (
-        <div style={{ height: '100%', textAlign: 'center', background: 'var(--bg-base)', borderRadius: 8, padding: 12, display: 'grid', placeItems: 'center' }}>
-          <img src={rawUrl} alt={name} style={{ maxWidth: '100%', maxHeight: '100%', objectFit: 'contain' }} />
-        </div>
-      ) : isPdf ? (
-        previewShell(t('file.pdfPreview'), <iframe title={name} src={rawUrl} style={{ width: '100%', height: '100%', border: 0, background: '#fff' }} />)
-      ) : isDocxPreview ? (
-        previewShell(t('file.wordPreview'), <Suspense fallback={<Spin />}><DocxFilePreview src={rawUrl} name={name} downloadUrl={downloadUrl} /></Suspense>)
-      ) : isExcelPreview ? (
-        previewShell(t('file.excelPreview'), <Suspense fallback={<Spin />}><ExcelFilePreview src={rawUrl} name={name} downloadUrl={downloadUrl} /></Suspense>)
-      ) : isPptxPreview ? (
-        previewShell(t('file.pptPreview'), <Suspense fallback={<Spin />}><PptxFilePreview src={rawUrl} name={name} downloadUrl={downloadUrl} /></Suspense>)
-      ) : isOffice ? (
-        previewShell(t('file.officePreview'), (
-          <div style={{ height: '100%', display: 'flex', flexDirection: 'column' }}>
-            <div style={{ flex: 1, minHeight: 0 }}>
-              <OfficePreview name={name} previewUrl={previewUrl} rawUrl={rawUrl} downloadUrl={downloadUrl} downloadName={downloadName} height="100%" />
-            </div>
-            <div style={{ padding: '8px 10px', borderTop: '1px solid var(--border-subtle)', color: 'var(--text-dim)', fontSize: 12, display: 'flex', gap: 8, alignItems: 'center', flexWrap: 'wrap' }}>
-              <span>{t('file.officePreviewHelp')}</span>
-              <Button size="small" type="primary" href={downloadUrl} download={downloadName}>{t('file.downloadFile')}</Button>
-              <Button size="small" href={rawUrl} target="_blank">{t('file.openRaw')}</Button>
-            </div>
-          </div>
-        ))
-      ) : (
-        <>
-          {err && <div style={{ color: '#f85149' }}>{err}</div>}
-          {!data && !err && <div style={{ textAlign: 'center', padding: 30 }}><Spin /></div>}
-          {data && data.binary && (
-            <div style={{ color: 'var(--text-dim)' }}>{t('file.binaryCannotPreview', { size: fmtSize(data.size) })}<a href={rawUrl} target="_blank" rel="noreferrer" style={{ color: accent }}>{t('file.downloadOrOpenRaw')}</a></div>
-          )}
-          {data && !data.binary && (
-            <>
-              {isSheetText
-                ? csvTable(data.content, ext === 'tsv' ? '\t' : ',')
-                : isMd && (!source || forcePreview)
-                  ? <div style={{ height: previewHeight, overflow: 'auto', padding: forcePreview ? '0 8px' : undefined }}><Markdown accent={accent} resolveHref={resolvePreviewHref} onLinkClick={openPreviewLink}>{data.content}</Markdown></div>
-                  : isHtml && (!source || forcePreview)
-                    // HTML 预览：不在前端拼 DOM，直接 iframe 后端服务代理(/api/file/raw 以 text/html 直出)，
-                    // 脚本/样式按原样运行。key 绑 mtime → 文件被外部(cc/codex)改动时 iframe 自动重载。
-                    ? <iframe key={data.mtime} title={name} src={rawUrl} style={{ display: 'block', width: '100%', height: previewHeight, border: 0, background: '#fff' }} />
-                  : (
-                    // 文本/代码/JSON/Markdown(源码) → Monaco 编辑器（行号、语法高亮、可编辑；截断的大文件只读）。
-                    // tab 语境下全屏无边框，背景由 CodeEditor 统一成应用底色。
-                    <div style={{ height: previewHeight, border: tabbed ? 'none' : '1px solid var(--border-subtle)', borderRadius: tabbed ? 0 : 8, overflow: 'hidden' }}>
-                      <Suspense fallback={<div style={{ height: '100%', display: 'grid', placeItems: 'center' }}><Spin /></div>}>
-                        <CodeEditor value={draft} language={monacoLangOf(path)} dark={mode === 'dark'} readOnly={!editable} onChange={setDraft} onSave={() => saveRef.current()} />
-                      </Suspense>
-                    </div>
-                  )}
-              {data.truncated && <div style={{ color: '#d29922', fontSize: 12, marginTop: 6 }}>⚠ {t('file.truncatedLong')}</div>}
-            </>
-          )}
-        </>
-      )}
-      {onOpenAgent && (
-        <Modal open={agentPick} title={t('file.openInAgent')} footer={null} onCancel={() => setAgentPick(false)}>
-          <Space direction="vertical" size={12} style={{ width: '100%' }}>
-            <div style={{ color: 'var(--text-dim)', fontFamily: 'ui-monospace, monospace', wordBreak: 'break-all' }}>{path}</div>
-            <Space>
-              <Button type="primary" onClick={() => { setAgentPick(false); onOpenAgent('claude', path) }}>Claude Code</Button>
-              <Button onClick={() => { setAgentPick(false); onOpenAgent('codex', path) }}>Codex</Button>
-            </Space>
-          </Space>
-        </Modal>
-      )}
-    </>
-  )
-
-  if (inline) {
-    return (
-      <div style={{ height: '100%', minHeight: 0, display: 'flex', flexDirection: 'column', background: 'var(--bg-base)' }}>
-        {/* 编辑器 tab 语境(tabbed)：文件名/操作已由外层 tab 承担 → 去掉整条标题栏，编辑器全屏。保存用 Ctrl/Cmd+S。 */}
-        {!tabbed && <div style={{ padding: '9px 12px', borderBottom: '1px solid var(--border-subtle)' }}>{titleNode}</div>}
-        <div style={{ flex: 1, minHeight: 0, overflow: 'hidden', padding: tabbed ? 0 : 12, position: 'relative' }}>
-          <ErrorBoundary>{bodyNode}</ErrorBoundary>
-          {/* tab 语境无头部：markdown 右上角 VSCode 式预览按钮（切换预览 / 侧栏打开预览） */}
-          {tabbed && (isMd || isHtml) && !forcePreview && data && !data.binary && (
-            <div style={{ position: 'absolute', top: 6, right: 8, zIndex: 10, display: 'inline-flex', gap: 2, background: 'color-mix(in srgb, var(--bg-base) 82%, transparent)', borderRadius: 8, padding: 2 }}>
-              <Tooltip title={source ? t('file.preview') : t('file.source')} placement="bottom">
-                <Button type="text" size="small" onClick={() => setSource((s) => !s)} style={{ color: !source ? accent : 'var(--text-dim)', width: 28, height: 28, display: 'inline-flex', alignItems: 'center', justifyContent: 'center' }}><PreviewIcon /></Button>
-              </Tooltip>
-              {onPreviewToSide && (
-                <Tooltip title={t('file.previewToSide')} placement="bottom">
-                  <Button type="text" size="small" onClick={() => onPreviewToSide(path)} style={{ color: 'var(--text-dim)', width: 28, height: 28, display: 'inline-flex', alignItems: 'center', justifyContent: 'center' }}><PreviewSideIcon /></Button>
-                </Tooltip>
-              )}
-              {isHtml && (
-                <Tooltip title={t('file.openInNewTab')} placement="bottom">
-                  <Button type="text" size="small" href={rawUrl} target="_blank" rel="noreferrer" style={{ color: 'var(--text-dim)', width: 28, height: 28, display: 'inline-flex', alignItems: 'center', justifyContent: 'center' }}><ExternalLinkIcon /></Button>
-                </Tooltip>
-              )}
-            </div>
-          )}
-          {/* 文件被外部(cc/codex)改动、但本地有未保存改动 → 提示条 */}
-          {stale && dirty && (
-            <div style={{ position: 'absolute', top: 6, left: '50%', transform: 'translateX(-50%)', zIndex: 11, display: 'inline-flex', alignItems: 'center', gap: 8, padding: '4px 10px', borderRadius: 8, fontSize: 12, background: 'var(--bg-container)', border: '1px solid #d29922', color: 'var(--text-bright)', boxShadow: 'var(--elevated-shadow)' }}>
-              <span>⚠ {t('file.changedOnDisk')}</span>
-              <Button size="small" danger onClick={reloadFromDisk}>{t('file.reloadFromDisk')}</Button>
-            </div>
-          )}
-        </div>
-      </div>
-    )
-  }
-
-  return (
-    <Modal open onCancel={onClose} footer={null} width="min(900px,94vw)" title={titleNode}>
-      <ErrorBoundary>{bodyNode}</ErrorBoundary>
-    </Modal>
-  )
-}
-
-function cellStyle(head: boolean): React.CSSProperties {
-  return {
-    padding: '6px 8px', border: '1px solid var(--border-subtle)', textAlign: 'left',
-    background: head ? 'var(--bg-container)' : 'transparent', color: head ? 'var(--text-bright)' : 'var(--text-dim)',
-    maxWidth: 220, overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap',
-  }
-}
-
 function formatJSON(text: string): string {
   try { return JSON.stringify(JSON.parse(text), null, 2) } catch { return text }
 }
