+++ frontend/src/fileview/CodeView.tsx
@@ -1,0 +1,25 @@
+// 文本/代码/JSON/Markdown(源码) → Monaco 编辑器（行号、语法高亮、可编辑；截断的大文件只读）。
+// Monaco 很重 → 懒加载，不进首屏包。tab 语境下全屏无边框，背景由 CodeEditor 统一成应用底色。
+import { lazy, Suspense } from 'react'
+import { Spin } from 'antd'
+
+const CodeEditor = lazy(() => import('../CodeEditor'))
+
+export function CodeView({ value, language, dark, readOnly, tabbed, height, onChange, onSave }: {
+  value: string
+  language: string
+  dark: boolean
+  readOnly: boolean
+  tabbed?: boolean
+  height: string
+  onChange: (v: string) => void
+  onSave: () => void
+}) {
+  return (
+    <div style={{ height, border: tabbed ? 'none' : '1px solid var(--border-subtle)', borderRadius: tabbed ? 0 : 8, overflow: 'hidden' }}>
+      <Suspense fallback={<div style={{ height: '100%', display: 'grid', placeItems: 'center' }}><Spin /></div>}>
+        <CodeEditor value={value} language={language} dark={dark} readOnly={readOnly} onChange={onChange} onSave={onSave} />
+      </Suspense>
+    </div>
+  )
+}
