+++ frontend/src/fileview/index.ts
@@ -1,0 +1,2 @@
+// 文件展示组件对外入口。页面/面板引用「展示一个文件」的能力统一从这里拿。
+export { FileView, Viewer } from './FileView'
